refactor(components/items): create `GenreBadge`
