+++ app/components/items/genre/genre-badge.spec.tsx
@@ -1,0 +1,11 @@
+import { render } from '@testing-library/react'
+
+import { GenreBadge } from './genre-badge'
+
+describe('GenreBadge', () => {
+  test('should match snapshot', () => {
+    const view = render(<GenreBadge genre="Black Metal" />)
+
+    expect(view).toMatchSnapshot()
+  })
+})
