+++ app/components/items/genre/genre-badge.tsx
@@ -1,0 +1,7 @@
+import { GenreProps } from './props'
+
+import { Badge } from '@app/components/ui/badge'
+
+export function GenreBadge({ genre }: GenreProps) {
+  return <Badge className="text-primary-foreground/80">{genre}</Badge>
+}
