+++ app/components/items/genre/genre-chip.tsx
@@ -1,8 +1,6 @@
-export interface GenreChipProps {
-  genre: string
-}
+import { GenreProps } from './props'
 
-export function GenreChip({ genre }: GenreChipProps) {
+export function GenreChip({ genre }: GenreProps) {
   return (
     <div className="w-max p-2 rounded-xl whitespace-nowrap bg-neutral-700 text-white">
       {genre}
