+++ app/components/items/genre/props.ts
@@ -1,0 +1,3 @@
+export interface GenreProps {
+  genre: string
+}
