+++ app/components/items/index.ts
@@ -7,4 +7,4 @@ export * from './item-image.skeleton'
 export * from './list/items-list.skeleton'
 export * from './item-top-card.skeleton'
 export * from './relative-time'
-export * from './genre-chip'
+export * from './genre/genre-chip'
